pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ado-workflows"
version = "0.2.2"
description = "Azure DevOps workflow automation library — PR context, repository discovery, and SDK client wrappers"
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = { text = "MIT" }
authors = [{ name = "Jack Pines" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
]

dependencies = [
    "actionable-errors>=0.1.0",
    "azure-devops>=7.1.0b4",
    "azure-identity>=1.19,<2",
]

[tool.uv]
prerelease = "if-necessary-or-explicit"

[project.optional-dependencies]
dev = [
    "pyright>=1.1,<2",
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=6,<7",
    "ruff>=0.8,<1",
    "taskipy>=1.14,<2",
    "pre-commit>=4,<5",
]

[project.urls]
Repository = "https://github.com/grimlor/ado-workflows"

# ---------------------------------------------------------------------------
# Taskipy
# ---------------------------------------------------------------------------
[tool.taskipy.tasks]
lint   = "uv run ruff check --fix src/ tests/"
format = "uv run ruff format src/ tests/"
type   = "uv run pyright src/ tests/"
test   = "uv run pytest tests/ -v"
cov    = "uv run pytest tests/ --cov=ado_workflows --cov-report=term-missing"
check  = "task format && task lint && task type && task test"

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
preview = true
explicit-preview-rules = true
select = [
    "E",        # pycodestyle errors
    "W",        # pycodestyle warnings
    "F",        # pyflakes
    "I",        # isort
    "N",        # pep8-naming
    "UP",       # pyupgrade
    "B",        # flake8-bugbear
    "SIM",      # flake8-simplify
    "TCH",      # flake8-type-checking
    "RUF",      # ruff-specific rules
    "D",        # pydocstyle
    "PLC0415",  # import-outside-top-level
    "PLC2701",  # import-private-name (preview)
]
ignore = [
    "E501",   # line length (handled by formatter)
    "D212",   # multi-line-summary-first-line (conflicts with D213)
    "D203",   # one-blank-line-before-class (conflicts with D211)
]

[tool.ruff.lint.isort]
known-first-party = ["ado_workflows"]
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "PLC0415",  # import-outside-top-level — tests may import inside functions
    "PLC2701",  # import-private-name — testing internal state is intentional
    "D205",     # BDD Given/When/Then blocks aren't summary+body
    "D400",     # BDD steps don't end with periods
    "D415",     # same as D400
    "D401",     # fixture docstrings are descriptive, not imperative
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Pyright
# ---------------------------------------------------------------------------
[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
stubPath = "typings"
venvPath = "."
venv = ".venv"

# ---------------------------------------------------------------------------
# python-semantic-release
# ---------------------------------------------------------------------------
[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = []
major_on_zero = false
allow_zero_version = true
commit_message = "chore(release): {version}"
tag_format = "v{version}"
branch = "main"
build_command = ""
upload_to_pypi = false
changelog_file = "CHANGELOG.md"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["feat", "fix", "perf", "refactor", "docs", "style", "test", "build", "ci", "chore"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["*_test.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers -m 'not integration and not live'"
markers = [
    "integration: marks tests requiring external services",
    "live: marks tests requiring live network access",
]
